requirements.txt
amqp==5.2.0
asgiref==3.7.2
asttokens==2.4.1
attrs==23.2.0
billiard==4.2.0
boto3==1.34.27
botocore==1.34.27
celery==5.3.6
certifi==2023.11.17
charset-normalizer==3.3.2
click==8.1.7
click-didyoumean==0.3.0
click-plugins==1.1.1
click-repl==0.3.0
coverage==7.4.0
decorator==5.1.1
dj-database-url==2.1.0
Django==5.0.1
django-appconf==1.0.6
django-betterforms==2.0.0
django-celery-email==3.0.0
django-cors-headers==4.3.1
django-create-react-app==0.9.0
django-extended-choices==1.3.3
django-extensions==3.2.3
django-heroku==0.3.1
django-multi-form-view==2.0.1
django-storages==1.14.2
django-utils-six==2.0
django-webpack-loader==3.0.1
djangorestframework==3.14.0
executing==2.0.1
Faker==22.5.1
freezegun==1.4.0
geckodriver-autoinstaller==0.1.0
gunicorn==21.2.0
h11==0.14.0
idna==3.6
iniconfig==2.0.0
ipdb==0.13.13
ipython==8.20.0
jedi==0.19.1
jmespath==1.0.1
kombu==5.3.5
matplotlib-inline==0.1.6
outcome==1.3.0.post0
packaging==23.2
parso==0.8.3
pexpect==4.9.0
pillow==10.2.0
pluggy==1.3.0
prompt-toolkit==3.0.43
psycopg2==2.9.9
psycopg2-binary==2.9.9
ptyprocess==0.7.0
pure-eval==0.2.2
Pygments==2.17.2
PySocks==1.7.1
pytest==7.4.4
python-dateutil==2.8.2
python-dotenv==1.0.1
pytz==2023.3.post1
requests==2.31.0
s3transfer==0.10.0
selenium==4.17.2
sentry-sdk==1.39.2
setuptools==69.0.3
six==1.16.0
sniffio==1.3.0
sortedcontainers==2.4.0
sqlparse==0.4.4
stack-data==0.6.3
tchappui-webdrivers==0.3.7
traitlets==5.14.1
trio==0.24.0
trio-websocket==0.11.1
typing_extensions==4.9.0
tzdata==2023.4
urllib3==2.0.7
vine==5.1.0
wcwidth==0.2.13
whitenoise==6.6.0
wsproto==1.2.0
